Fixed support for config updates

+++ avorion-manager/UpdateManager.sh
@@ -8,30 +8,45 @@ if [ "${CurrentVersion}" != "${VERSION}" ]; then
   wget -O UpdateDirtyServerManager.tar.gz https://github.com/dirtyredz/Dirty-Server-Manager/releases/download/${VERSION}/DirtyServerManager.tar.gz
   echo Installing $VERSION
 
-  cp manager-config.ini manager-config.ini.old
-  cp avorion-manager/PHPConfig.ini avorion-manager/PHPConfig.ini.old
+  if [ -f manager-config.ini ]; then
+    cp manager-config.ini manager-config.ini.old
+    cp manager-config.ini manager-config.ini.backup
+  fi
+  if [ -f avorion-manager/PHPConfig.ini ]; then
+    cp avorion-manager/PHPConfig.ini avorion-manager/PHPConfig.ini.old
+    cp avorion-manager/PHPConfig.ini avorion-manager/PHPConfig.ini.backup
+  fi
 
   tar -xvf UpdateDirtyServerManager.tar.gz --exclude='Database.php' --exclude='avorion-manager/PlayerData.php' --exclude='avorion-manager/SectorData.php' --exclude='avorion-manager/logs'
 
-  input="manager-config.ini.old"
-  while IFS= read -r var
-  do
-    OldLineConfigName=$(echo $var | sed -e 's/=.*//g')
-    sed -i "s/${OldLineConfigName}.*/${var}/" manager-config.ini
-  done < "$input"
-  rm manager-config.ini.old
+  if [ -f manager-config.ini.old ]; then
+    sed -i 's/^;.*//g' manager-config.ini.old
+    input="manager-config.ini.old"
+    while IFS= read -r var
+    do
+      if [ ! -z "$var" ]; then
+        OldLineConfigName=$(echo $var | sed -e '0,/=.*/s///')
+        sed -i "s~${OldLineConfigName}.*~${var}~" manager-config.ini
+      fi
+    done < "$input"
+    rm manager-config.ini.old
+  fi
 
-  input="avorion-manager/PHPConfig.ini.old"
-  while IFS= read -r var
-  do
-    OldLineConfigName=$(echo $var | sed -e 's/=.*//g')
-    sed -i "s/${OldLineConfigName}.*/${var}/" avorion-manager/PHPConfig.ini
-  done < "$input"
-  rm avorion-manager/PHPConfig.ini.old
+  if [ -f avorion-manager/PHPConfig.ini.old ]; then
+    sed -i 's/^;.*//g' avorion-manager/PHPConfig.ini.old
+    input="avorion-manager/PHPConfig.ini.old"
+    while IFS= read -r var
+    do
+      if [ ! -z "$var" ]; then
+        OldLineConfigName=$(echo $var | sed -e '0,/=.*/s///')
+        sed -i "s~${OldLineConfigName}.*~${var}~" avorion-manager/PHPConfig.ini
+      fi
+    done < "$input"
+    rm avorion-manager/PHPConfig.ini.old
+  fi
 
   rm UpdateDirtyServerManager.tar.gz
   echo Update complete.
 else
   echo "Current Version: ${CurrentVersion}, Dirty Server Manager is Up To Date!"
-
 fi
